refactor(media): extract RTCConnection from WebRTCTransport
