+++ src/modules/media/WebRTC.ts
@@ -1,54 +1,45 @@
 import type { CallStats } from "@/modules/call/Stats";
-import { RTCStatsAdapter } from "@/modules/media/composition";
-import {
-    type ConnectivityIssue,
-    DEFAULT_ICE_GATHERING_TIMEOUT_MS,
-    DEFAULT_ICE_SERVERS,
-    type IceCandidateKind,
-    type IceConfig,
-    type IceDiagnostics,
-} from "@/modules/media/ICEDiagnostics";
+import { RTCConnection, RTCStatsAdapter } from "@/modules/media/composition";
+import type { ConnectivityIssue, IceConfig, IceDiagnostics } from "@/modules/media/ICEDiagnostics";
 import type { Events, ITransport, TransportStatus } from "@/modules/media/ITransport";
 import type { MediaManager } from "@/modules/media/MediaManager";
 import { EventEmitter } from "@/modules/shared/EventEmitter";
 
-const SYMMETRIC_NAT_DETECTION_WINDOW_MS = 10_000;
-
 export class WebRTCTransport extends EventEmitter<Events> implements ITransport {
     readonly kind = "webrtc" as const;
-    status: TransportStatus = "disconnected";
     peerMuted = false;
     audioAnalyser: Promise<AnalyserNode>;
 
-    get stats(): CallStats {
-        return this.statsAdapter.snapshot();
+    private readonly connection: RTCConnection;
+    private readonly statsAdapter: RTCStatsAdapter;
+    private readonly hasRemoteOffer: boolean;
+    private statsJob = 0;
+    private tracksAdded = false;
+    private startedOnce = false;
+    private stoppedOnce = false;
+
+    get status(): TransportStatus {
+        return this.connection.status;
     }
 
-    readonly answer: Promise<RTCSessionDescriptionInit>;
+    get pc(): RTCPeerConnection {
+        return this.connection.pc;
+    }
 
-    private pc: RTCPeerConnection;
-    private statsAdapter: RTCStatsAdapter;
-    private remoteOffer?: RTCSessionDescriptionInit;
-    private answerResolver: PromiseWithResolvers<RTCSessionDescriptionInit>;
-    private statsJob = 0;
-    private started = false;
-    private offerCreated = false;
-    private stopped = false;
+    get answer(): Promise<RTCSessionDescriptionInit> {
+        return this.connection.answer;
+    }
 
-    private readonly gatheringTimeoutMs: number;
-    private gatheringStartedAt = 0;
-    private candidatesByType: Record<IceCandidateKind, number> = {
-        host: 0,
-        srflx: 0,
-        prflx: 0,
-        relay: 0,
-    };
-    private symmetricNatTimer = 0;
-    private _emittedConnectivityIssues = new Set<ConnectivityIssue>();
-    lastDiagnostics: IceDiagnostics | null = null;
+    get lastDiagnostics(): IceDiagnostics | null {
+        return this.connection.lastDiagnostics;
+    }
 
     get emittedConnectivityIssues(): ReadonlySet<ConnectivityIssue> {
-        return this._emittedConnectivityIssues;
+        return this.connection.emittedConnectivityIssues;
+    }
+
+    get stats(): CallStats {
+        return this.statsAdapter.snapshot();
     }
 
     constructor(
@@ -58,35 +49,28 @@ export class WebRTCTransport extends EventEmitter<Events> implements ITransport 
     ) {
         super();
 
-        this.gatheringTimeoutMs = iceConfig?.gatheringTimeoutMs ?? DEFAULT_ICE_GATHERING_TIMEOUT_MS;
-        const iceServers = iceConfig?.iceServers ?? DEFAULT_ICE_SERVERS;
+        this.hasRemoteOffer = !!offer;
+        this.connection = new RTCConnection(offer, iceConfig);
+        this.statsAdapter = new RTCStatsAdapter(this.connection.pc, this.mediaManager.audioContext);
 
-        this.pc = new RTCPeerConnection({ iceServers });
-        this.statsAdapter = new RTCStatsAdapter(this.pc, this.mediaManager.audioContext);
-        if (offer) this.remoteOffer = { type: "offer", sdp: offer };
+        this.connection.on("statusChanged", (s) => {
+            this.emit("statusChanged", s);
+            // Autonomous close (pc.connectionState transitions to "closed" outside
+            // of stop()) still needs to release the mic. RTCConnection owns the
+            // peer-connection lifecycle but has no MediaManager dependency.
+            if (this.connection.pc.connectionState === "closed") void this.mediaManager.stopMedia();
+        });
+        this.connection.on("iceDiagnostics", (d) => this.emit("iceDiagnostics", d));
+        this.connection.on("connectivityIssue", (i) => this.emit("connectivityIssue", i));
 
         const { promise: audioAnalyserPromise, resolve: resolveAudioAnalyser } = Promise.withResolvers<AnalyserNode>();
         this.audioAnalyser = audioAnalyserPromise;
 
-        this.answerResolver = Promise.withResolvers<RTCSessionDescriptionInit>();
-        this.answer = this.answerResolver.promise;
-
-        this.pc.onicecandidate = (event) => {
-            const candidate = event.candidate;
-            if (!candidate) return;
-            const kind = candidate.type as IceCandidateKind | undefined;
-            if (kind && kind in this.candidatesByType) this.candidatesByType[kind] += 1;
-        };
-
-        this.pc.oniceconnectionstatechange = () => {
-            if (this.pc.iceConnectionState === "failed") {
-                this.emitIssue("ICE_CONNECTION_FAILED");
-            }
-            if (this.pc.iceConnectionState === "connected" || this.pc.iceConnectionState === "completed") {
-                clearTimeout(this.symmetricNatTimer);
-            }
-        };
-
+        // Audio pipe wiring stays here until C7. `ontrack` fires once when the
+        // remote stream lands — extract a MediaStream-backed analyser so the
+        // peerMute/peerUnmute events and the analyser promise both have a source.
+        // The muted <audio> element exists as a Chromium workaround
+        // (issues.chromium.org/issues/40094084) to keep audio flowing.
         this.pc.ontrack = (event) => {
             const remoteStream = event.streams[0];
 
@@ -118,144 +102,48 @@ export class WebRTCTransport extends EventEmitter<Events> implements ITransport 
             resolveAudioAnalyser(analyser);
         };
 
-        this.pc.onconnectionstatechange = () => {
-            if (this.pc.connectionState === "connecting") this.setStatus("connecting");
-            if (this.pc.connectionState === "disconnected" || this.pc.connectionState === "closed") {
-                this.setStatus("disconnected");
-            }
-            if (this.pc.connectionState === "closed") this.mediaManager.stopMedia();
-            if (this.pc.connectionState === "connected") this.setStatus("connected");
-        };
     }
 
     async start(): Promise<void> {
-        if (this.started) return;
-        this.started = true;
+        if (this.startedOnce) return;
+        this.startedOnce = true;
 
-        if (this.remoteOffer) {
-            const micStream = await this.mediaManager.startMedia();
-
-            for (const track of micStream.getTracks()) {
-                track.enabled = !this.mediaManager.muted;
-                this.pc.addTrack(track, micStream);
-            }
-
-            await this.pc.setRemoteDescription(this.remoteOffer);
-            const answer = await this.pc.createAnswer();
-            await this.pc.setLocalDescription(answer);
-
-            await this.waitForIceGathering();
-
-            this.answerResolver.resolve(this.pc.localDescription as RTCSessionDescription);
-        }
+        if (this.hasRemoteOffer) await this.attachMicTracks();
+        await this.connection.start();
 
         await this.tickStats();
         this.statsJob = setInterval(() => void this.tickStats(), 200) as unknown as number;
     }
 
-    private async tickStats(): Promise<void> {
-        await this.statsAdapter.refresh();
-        this.emit("statsChanged", this.statsAdapter.snapshot());
+    async createOffer(): Promise<string> {
+        await this.attachMicTracks();
+        return this.connection.createOffer();
     }
 
-    async createOffer(): Promise<string> {
-        if (this.offerCreated) return this.pc.localDescription?.sdp as string;
-        this.offerCreated = true;
+    async setAnswer(sdp: string): Promise<void> {
+        await this.connection.setAnswer(sdp);
+    }
 
+    async stop(): Promise<void> {
+        if (this.stoppedOnce) return;
+        this.stoppedOnce = true;
+        clearInterval(this.statsJob);
+        await this.connection.stop();
+        await this.mediaManager.stopMedia();
+    }
+
+    private async attachMicTracks(): Promise<void> {
+        if (this.tracksAdded) return;
+        this.tracksAdded = true;
         const micStream = await this.mediaManager.startMedia();
-
         for (const track of micStream.getTracks()) {
             track.enabled = !this.mediaManager.muted;
             this.pc.addTrack(track, micStream);
         }
-
-        const offer = await this.pc.createOffer();
-        await this.pc.setLocalDescription(offer);
-
-        await this.waitForIceGathering();
-
-        return this.pc.localDescription?.sdp as string;
     }
 
-    async setAnswer(sdp: string): Promise<void> {
-        await this.pc.setRemoteDescription({ type: "answer", sdp });
-    }
-
-    private async waitForIceGathering(): Promise<void> {
-        this.gatheringStartedAt = Date.now();
-
-        const timedOut = await this.raceGatheringWithTimeout();
-
-        const duration = Date.now() - this.gatheringStartedAt;
-        const stunReached = this.candidatesByType.srflx > 0;
-        const turnReached = this.candidatesByType.relay > 0;
-
-        const diag: IceDiagnostics = {
-            gatheringDurationMs: duration,
-            gatheringTimedOut: timedOut,
-            candidatesByType: { ...this.candidatesByType },
-            stunReached,
-            turnReached,
-        };
-        this.lastDiagnostics = diag;
-        this.emit("iceDiagnostics", diag);
-
-        if (timedOut) this.emitIssue("ICE_GATHERING_TIMEOUT");
-        if (timedOut && !stunReached) this.emitIssue("STUN_UNREACHABLE");
-        if (this.candidatesByType.host === 0) this.emitIssue("NO_HOST_CANDIDATES");
-
-        this.scheduleSymmetricNatCheck(stunReached);
-    }
-
-    private raceGatheringWithTimeout(): Promise<boolean> {
-        if (this.pc.iceGatheringState === "complete") return Promise.resolve(false);
-
-        return new Promise<boolean>((resolve) => {
-            const handler = () => {
-                if (this.pc.iceGatheringState !== "complete") return;
-                this.pc.removeEventListener("icegatheringstatechange", handler);
-                clearTimeout(timer);
-                resolve(false);
-            };
-
-            const timer = setTimeout(() => {
-                this.pc.removeEventListener("icegatheringstatechange", handler);
-                resolve(true);
-            }, this.gatheringTimeoutMs);
-
-            this.pc.addEventListener("icegatheringstatechange", handler);
-        });
-    }
-
-    private scheduleSymmetricNatCheck(stunReached: boolean) {
-        if (!stunReached) return;
-        if (this.symmetricNatTimer) return;
-        this.symmetricNatTimer = setTimeout(() => {
-            const noConnection =
-                this.pc.iceConnectionState !== "connected" && this.pc.iceConnectionState !== "completed";
-            if (noConnection) this.emitIssue("SYMMETRIC_NAT_SUSPECTED");
-        }, SYMMETRIC_NAT_DETECTION_WINDOW_MS) as unknown as number;
-    }
-
-    private emitIssue(issue: ConnectivityIssue) {
-        if (this._emittedConnectivityIssues.has(issue)) return;
-        this._emittedConnectivityIssues.add(issue);
-        this.emit("connectivityIssue", issue);
-    }
-
-    async stop(): Promise<void> {
-        if (this.stopped) return;
-        this.stopped = true;
-
-        clearInterval(this.statsJob);
-        clearTimeout(this.symmetricNatTimer);
-
-        this.pc.close();
-        await this.mediaManager.stopMedia();
-    }
-
-    private setStatus(status: TransportStatus) {
-        this.status = status;
-        this.emit("statusChanged", status);
+    private async tickStats(): Promise<void> {
+        await this.statsAdapter.refresh();
+        this.emit("statsChanged", this.statsAdapter.snapshot());
     }
 }
